Do not attempt to install gems if there is no Gemfile in InstallRhizome
This is another preparation for the new services. Not every service would need
gems and they wouldn't have their own Gemfiles. The primary goal here is making
BootstrapRhizome and InstallRhizome useable, not just for host but also for VMs
created by other services.

+++ prog/install_rhizome.rb
@@ -37,7 +37,9 @@ class Prog::InstallRhizome < Prog::Base
 
   label def install_gems
     sshable.cmd("bundle install --gemfile common/Gemfile --path common/vendor/bundle")
-    sshable.cmd("bundle install --gemfile #{frame["target_folder"]}/Gemfile --path #{frame["target_folder"]}/vendor/bundle")
+    if File.exist?("#{Config.root}/rhizome/#{frame["target_folder"]}/Gemfile")
+      sshable.cmd("bundle install --gemfile #{frame["target_folder"]}/Gemfile --path #{frame["target_folder"]}/vendor/bundle")
+    end
     pop "installed rhizome"
   end
 end
